+++ src/SwaggerProvider.DesignTime/SchemaParser.fs
@@ -8,6 +8,8 @@ module Parser =
     // Type that hold parsing context to resolve `$ref`s
     type ParserContext =
         {
+            /// An object to hold type definitions
+            Definitions: Map<string, SchemaObject>
             /// An object to hold parameters that can be used across operations
             Parameters: Map<string, ParameterObject>
             /// An object to hold responses that can be used across operations.
@@ -44,6 +46,7 @@ module Parser =
         /// Default empty context
         static member Empty =
             {
+                Definitions = Map.empty<_,_>
                 Parameters = Map.empty<_,_>
                 Responses = Map.empty<_,_>
                 ApplicableParameters = [||]
@@ -225,7 +228,7 @@ module Parser =
                 Description = obj.GetRequiredField("description", spec).AsString()
                 Schema =
                     obj.TryGetProperty("schema")
-                    |> Option.map (parseSchemaObject Map.empty)
+                    |> Option.map (parseSchemaObject context.Definitions)
             }
 
     /// Parses the JsonValue as a Responses  Definition Object
@@ -328,7 +331,7 @@ module Parser =
         |> Array.concat
 
     /// Parse Definitions Object as a SchemaObject[]
-    let parseDefinitionsObject (obj:SchemaNode) : (string*SchemaObject)[] =
+    let parseDefinitionsObject (obj:SchemaNode) : Map<string,SchemaObject> =
         obj.Properties()
         |> Array.fold (fun tys (name, schemaObj) ->
             Map.add
@@ -336,7 +339,6 @@ module Parser =
                 (parseSchemaObject tys schemaObj)
                 tys
             ) Map.empty<_,_>
-        |> Map.toArray
 
     /// Parses the JsonValue as an InfoObject.
     let parseInfoObject (obj:SchemaNode) : InfoObject =
@@ -367,6 +369,10 @@ module Parser =
         let context =
             {
                 ParserContext.Empty with
+                    Definitions =
+                        match obj.TryGetProperty("definitions") with
+                        | None -> Map.empty<_,_>
+                        | Some(definitions) -> parseDefinitionsObject definitions
                     Parameters =
                         match obj.TryGetProperty("parameters") with
                         | None -> Map.empty<_,_>
@@ -390,8 +396,5 @@ module Parser =
             Paths =
                 obj.GetRequiredField("paths", spec)
                 |> (parsePathsObject context)
-            Definitions =
-                match obj.TryGetProperty("definitions") with
-                | None -> [||]
-                | Some(definitions) -> parseDefinitionsObject definitions
+            Definitions = context.Definitions |> Map.toArray
         }